+++ kwimage/structs/boxes.py
@@ -3172,7 +3172,7 @@ class Boxes(_BoxConversionMixins, _BoxPropertyMixins, _BoxTransformMixins,
                 (must be of same length)
 
         Returns:
-            Boxes: the intersection geometry
+            Boxes: the component-wise intersection geometry
 
         Examples:
             >>> # xdoctest: +IGNORE_WHITESPACE
@@ -3198,7 +3198,10 @@ class Boxes(_BoxConversionMixins, _BoxPropertyMixins, _BoxTransformMixins,
         is_bad = np.any(tl > br, axis=1)
         ltrb = np.concatenate([tl, br], axis=-1)
 
-        ltrb[is_bad] = np.nan
+        if is_bad.any():
+            if ltrb.dtype.kind != 'f':
+                ltrb = ltrb.to(float)
+            ltrb[is_bad] = np.nan
 
         isect = Boxes(ltrb, 'ltrb')
 
@@ -3206,16 +3209,21 @@ class Boxes(_BoxConversionMixins, _BoxPropertyMixins, _BoxTransformMixins,
 
     def union_hull(self, other):
         """
-        Componentwise hull union between two sets of Boxes
+        Componentwise bounds of the union between two sets of Boxes
 
         NOTE: convert to polygon to do a real union.
+
+        Note:
+            This is not a real hull. A better name for this might be
+            union_bounds because we are returning the axis-aligned bounds of
+            the convex hull of the union.
 
         Args:
             other (Boxes): boxes to union with this object.
                 (must be of same length)
 
         Returns:
-            Boxes: unioned boxes
+            Boxes: bounding box of the unioned boxes
 
         Examples:
             >>> # xdoctest: +IGNORE_WHITESPACE
@@ -3238,7 +3246,10 @@ class Boxes(_BoxConversionMixins, _BoxPropertyMixins, _BoxTransformMixins,
         is_bad = np.any(tl > br, axis=1)
         ltrb = np.concatenate([tl, br], axis=-1)
 
-        ltrb[is_bad] = np.nan
+        if is_bad.any():
+            if ltrb.dtype.kind != 'f':
+                ltrb = ltrb.to(float)
+            ltrb[is_bad] = np.nan
 
         isect = Boxes(ltrb, 'ltrb')