+++ kwimage/structs/polygon.py
@@ -1689,6 +1689,10 @@ class Polygon(_generic.Spatial, _PolyArrayBackend, _PolyWarpMixin, _ShapelyMixin
         Returns:
             kwimage.Boxes
         """
+        ub.schedule_deprecation(
+            'kwimage', 'Polygon.bounding_box', 'function',
+            migration='Use the box method instead.', deprecate='0.9.20',
+            error='1.0.0', remove='1.1.0')
         import kwimage
         xys = self.data['exterior'].data
         lt = xys.min(axis=0)
@@ -1719,8 +1723,8 @@ class Polygon(_generic.Spatial, _PolyArrayBackend, _PolyWarpMixin, _ShapelyMixin
         lt = xys.min(axis=0)
         rb = xys.max(axis=0)
         ltrb = np.hstack([lt, rb])
-        boxes = kwimage.Box.coerce(ltrb, format='ltrb')
-        return boxes
+        box = kwimage.Box.coerce(ltrb, format='ltrb')
+        return box
 
     def bounding_box_polygon(self):
         """
@@ -1793,7 +1797,7 @@ class Polygon(_generic.Spatial, _PolyArrayBackend, _PolyWarpMixin, _ShapelyMixin
         Example:
             >>> # xdoctest: +REQUIRES(module:rasterio)
             >>> import kwimage
-            >>> mask = kwimage.Mask.random()
+            >>> mask = kwimage.Mask.random(rng=0)
             >>> self = mask.to_multi_polygon(pixels_are='areas').data[0]
             >>> image = np.zeros_like(mask.data)
             >>> self.fill(image, pixels_are='areas')
@@ -2140,19 +2144,19 @@ class Polygon(_generic.Spatial, _PolyArrayBackend, _PolyWarpMixin, _ShapelyMixin
             edgecolor (None | Any):
                 if None, uses the value of ``color``.
                 Otherwise the color of the border when linewidth > 0.
-                Extended types Coercable[kwimage.Color].
+                Extended types Coercible[kwimage.Color].
 
             facecolor (None | Any):
                 if None, uses the value of ``color``.
                 Otherwise, color of the border when fill=True.
-                Extended types Coercable[kwimage.Color].
+                Extended types Coercible[kwimage.Color].
 
             vertex (float):
                 if non-zero, draws vertexes on the polygon with this radius.
 
             vertexcolor (Any):
                 color of vertexes
-                Extended types Coercable[kwimage.Color].
+                Extended types Coercible[kwimage.Color].
 
         Returns:
             matplotlib.patches.PathPatch | None :
@@ -2564,6 +2568,8 @@ class MultiPolygon(_generic.ObjectList, _ShapelyMixin):
         """
         Return the bounding box of the multi polygon
 
+        DEPRECATED: Use singular :func:`box` instead.
+
         Returns:
             kwimage.Boxes: a Boxes object with one box that encloses all
                 polygons
@@ -2577,6 +2583,10 @@ class MultiPolygon(_generic.ObjectList, _ShapelyMixin):
             >>> areas2 = np.array([s.area[0] for s in sub_boxes])
             >>> assert np.allclose(areas1, areas2)
         """
+        ub.schedule_deprecation(
+            'kwimage', 'MultiPolygon.bounding_box', 'function',
+            migration='Use the box method instead.', deprecate='0.9.20',
+            error='1.0.0', remove='1.1.0')
         import kwimage
         lt = np.array([np.inf, np.inf])
         rb = np.array([-np.inf, -np.inf])
@@ -2587,6 +2597,33 @@ class MultiPolygon(_generic.ObjectList, _ShapelyMixin):
         ltrb = np.hstack([lt, rb])[None, :]
         boxes = kwimage.Boxes(ltrb, 'ltrb')
         return boxes
+
+    def box(self):
+        """
+        Returns an axis-aligned bounding box for the segmentation
+
+        Returns:
+            kwimage.Box
+
+        Example:
+            >>> from kwimage.structs.polygon import *  # NOQA
+            >>> self = MultiPolygon.random(rng=0, n=10)
+            >>> boxes = self.box()
+            >>> sub_boxes = [d.box() for d in self.data]
+            >>> areas1 = np.array([s.intersection(boxes).area for s in sub_boxes])
+            >>> areas2 = np.array([s.area for s in sub_boxes])
+            >>> assert np.allclose(areas1, areas2)
+        """
+        import kwimage
+        lt = np.array([np.inf, np.inf])
+        rb = np.array([-np.inf, -np.inf])
+        for data in self.data:
+            xys = data.data['exterior'].data
+            lt = np.minimum(lt, xys.min(axis=0))
+            rb = np.maximum(rb, xys.max(axis=0))
+        ltrb = np.hstack([lt, rb])[None, :]
+        box = kwimage.Box.coerce(ltrb, format='ltrb')
+        return box
 
     def to_mask(self, dims=None, pixels_are='points'):
         """